backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# PyTorch y dependencias de ML
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
scipy==1.11.3

# Procesamiento de audio
soundfile==0.12.1
librosa==0.10.1

# Procesamiento de texto
g2p-en==2.1.0
pypinyin==0.49.0
unidecode==1.3.7

# Visualización y utilidades
matplotlib==3.8.2
pyyaml==6.0.1
tqdm==4.66.1
numba==0.60.0
inflect==7.0.0
tensorboard==2.15.1
tgt==1.4.4
pyworld==0.3.4
scikit-learn==1.4.0